# api/room_handler.go
package api

import (
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/raminfathi/GoTel/db"
	"github.com/raminfathi/GoTel/types"

	"github.com/gofiber/fiber/v3"
	"go.mongodb.org/mongo-driver/v2/bson"
)

type RoomHandler struct {
	store *db.Store
}

func NewRoomHandler(store *db.Store) *RoomHandler {
	return &RoomHandler{
		store: store,
	}
}

// HandleGetRooms returns all rooms
// @Summary      Get all rooms
// @Description  Get a list of all rooms
// @Tags         room
// @Accept       json
// @Produce      json
// @Param        X-Api-Token header string true "Token"
// @Success      200  {array}   types.Room
// @Router       /room [get]
func (h *RoomHandler) HandleGetRooms(c fiber.Ctx) error {

	cacheKey := "rooms-" + c.OriginalURL()
	val, err := h.store.Cache.Get(c.Context(), cacheKey)
	if err == nil && val != "" {
		var rooms []*types.Room
		if err := json.Unmarshal([]byte(val), &rooms); err == nil {
			return c.JSON(rooms)
		}
	}
	filter := bson.M{}
	if hotelID := c.Query("hotelId"); hotelID != "" {
		oid, err := bson.ObjectIDFromHex(hotelID)
		if err != nil {
			return types.ErrInvalidID()
		}
		filter["hotelID"] = oid
	}
	rooms, err := h.store.Room.GetRooms(c.Context(), filter)
	if err != nil {
		return types.ErrResourceNotFound("rooms")
	}
	serialized, err := json.Marshal(rooms)
	if err == nil {
		h.store.Cache.Set(c.Context(), cacheKey, serialized, time.Minute*1)
	}

	return c.JSON(rooms)
}

// HandleBookRoom creates a booking for a room
// @Summary      Book a room
// @Description  Book a specific room for a date range
// @Tags         room
// @Accept       json
// @Produce      json
// @Param        id      path    string                true  "Room ID"
// @Param        request body    types.BookRoomParams  true  "Booking Details"
// @Param        X-Api-Token header string true "Token"
// @Success      200     {object}  types.Booking
// @Failure      400     {object}  map[string]string
// @Router       /room/{id}/book [post]
func (h *RoomHandler) HandleBookRoom(c fiber.Ctx) error {
	var params types.BookRoomParams
	if err := c.Bind().Body(&params); err != nil {
		return err
	}

	if errors := ValidateRequest(params); errors != nil {
		return c.Status(fiber.StatusBadRequest).JSON(errors)
	}
	roomID, err := bson.ObjectIDFromHex(c.Params("id"))
	if err != nil {
		return types.ErrInvalidID()
	}
	user, ok := c.Locals("user").(*types.User)
	if !ok {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "user not found in context"})
	}
	ok, err = h.store.Booking.IsRoomAvailable(c.Context(), roomID, params.FromDate, params.TillDate)
	if err != nil {
		return err
	}
	if !ok {
		return c.Status(http.StatusBadRequest).JSON(genericResp{
			Type: "error",
			Msg:  fmt.Sprintf("room %s already booked", c.Params("id")),
		})
	}
	booking := types.Booking{
		UserID:     user.ID,
		RoomID:     roomID,
		FromDate:   params.FromDate,
		TillDate:   params.TillDate,
		NumPersons: params.NumPersons,
	}

	inserted, err := h.store.Booking.InsertBooking(c.Context(), &booking)
	if err != nil {
		return err
	}
	return c.JSON(inserted)
}

// HandlePostRoom adds a new room (Admin only)
// @Summary      Add a room
// @Description  Add a new room to a hotel
// @Tags         admin
// @Accept       json
// @Produce      json
// @Param        request body types.CreateRoomParams true "Room Data"
// @Param        X-Api-Token header string true "Token"
// @Success      200  {object}  types.Room
// @Router       /admin/room [post]
func (h *RoomHandler) HandlePostRoom(c fiber.Ctx) error {
	var params types.CreateRoomParams
	if err := c.Bind().Body(&params); err != nil {
		return types.ErrBadRequest()
	}

	if errors := params.Validate(); len(errors) > 0 {
		return c.Status(fiber.StatusBadRequest).JSON(errors)
	}

	hotelOID, err := bson.ObjectIDFromHex(params.HotelID)
	if err != nil {
		return types.ErrInvalidID()
	}

	room := types.Room{
		HotelID:   hotelOID,
		BasePrice: params.BasePrice,
		Price:     params.BasePrice,
	}

	switch params.Type {
	case "single":
		room.Type = types.Single
	case "double":
		room.Type = types.Double
	case "seaview":
		room.Type = types.SeaView
	default:
		room.Type = types.Single
	}

	insertedRoom, err := h.store.Room.InsertRoom(c.Context(), &room)
	if err != nil {
		return err
	}

	return c.JSON(insertedRoom)
}